backend/requirements.txt
fastapi
uvicorn[standard]
neo4j>=5.0
sentence-transformers>=2.2.0
groq>=0.4.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
mysql-connector-python>=8.0.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0
pydantic>=2.0.0
python-multipart>=0.0.6


